made jitter a double and moved init of WindowSize to after we determine geometry from commands

+++ exercises/exercise_4/question_5/inc/common.h
@@ -45,7 +45,7 @@ struct threadConfig_s
 
 struct TransformAnalysis_s
 {
-   uint32_t jitter;
+   double jitter;
    uint32_t deadline_missed;
 };
